CMM: Eliminate the risk of cyclic references
Replace all pointers in MuxItem_t other than pin objects with self-owned buffers.
If MuxItem_t is embedded in other objects which are allocated with GC, these pointers inside MuxItem_t can point to other parts within CMM which in case points back to the objects that embed MuxItem_t instances, this leads to cyclic referenes, and GC is proved to can be fooled to leak memory.